chore(key-management): refactor shared staking key util to key management package LW-6803
