--- packages/hardware-ledger/src/transformers/keyPaths.ts
@@ -1,17 +1,0 @@
-import * as Ledger from '@cardano-foundation/ledgerjs-hw-app-cardano';
-import { CardanoKeyConst, GroupedAddress, util } from '@cardano-sdk/key-management';
-
-export const stakeKeyPathFromGroupedAddress = (address: GroupedAddress | undefined): Ledger.BIP32Path | null => {
-  if (!address) return null;
-  if (address && address.stakeKeyDerivationPath) {
-    return [
-      util.harden(CardanoKeyConst.PURPOSE),
-      util.harden(CardanoKeyConst.COIN_TYPE),
-      util.harden(address.accountIndex),
-      address.stakeKeyDerivationPath.role,
-      address.stakeKeyDerivationPath.index
-    ];
-  }
-
-  return null;
-};
