+++ packages/hardware-ledger/src/transformers/txOut.ts
@@ -3,7 +3,6 @@ import { CML, Cardano, Serialization, coreToCml } from '@cardano-sdk/core';
 import { HexBlob, InvalidArgumentError, ManagedFreeableScope, Transform, usingAutoFree } from '@cardano-sdk/util';
 import { LedgerTxTransformerContext } from '../types';
 import { mapTokenMap } from './assets';
-import { stakeKeyPathFromGroupedAddress } from './keyPaths';
 import { util } from '@cardano-sdk/key-management';
 
 const toInlineDatum: Transform<Cardano.Datum, Ledger.Datum> = (datum) => ({
@@ -24,7 +23,7 @@ const toDestination: Transform<Cardano.TxOut, Ledger.TxOutputDestination, Ledger
 
   if (knownAddress) {
     const paymentKeyPath = util.paymentKeyPathFromGroupedAddress(knownAddress);
-    const stakeKeyPath = stakeKeyPathFromGroupedAddress(knownAddress);
+    const stakeKeyPath = util.stakeKeyPathFromGroupedAddress(knownAddress);
 
     if (!stakeKeyPath) throw new InvalidArgumentError('txOut', 'Missing stake key key path.');
