+++ packages/hardware-trezor/src/transformers/txOut.ts
@@ -4,7 +4,6 @@ import { GroupedAddress, util } from '@cardano-sdk/key-management';
 import { InvalidArgumentError, Transform } from '@cardano-sdk/util';
 import { TrezorTxOutputDestination, TrezorTxTransformerContext } from '../types';
 import { mapTokenMap } from './assets';
-import { stakeKeyPathFromGroupedAddress } from './keyPaths';
 
 const toDestination: Transform<Cardano.TxOut, TrezorTxOutputDestination, TrezorTxTransformerContext> = (
   txOut,
@@ -19,7 +18,7 @@ const toDestination: Transform<Cardano.TxOut, TrezorTxOutputDestination, TrezorT
   }
 
   const paymentPath = util.paymentKeyPathFromGroupedAddress(knownAddress);
-  const stakingPath = stakeKeyPathFromGroupedAddress(knownAddress);
+  const stakingPath = util.stakeKeyPathFromGroupedAddress(knownAddress);
 
   if (!stakingPath) throw new InvalidArgumentError('txOut', 'Missing staking key key path.');
